Save Product split before upgrade
