+++ VSKeyExtractor.slnx
@@ -1,3 +1,3 @@
 <Solution>
-  <Project Path="VSKeyExtractor.csproj" />
+  <Project Path="VSKeyExtractor.csproj" Id="ccc458e3-ba11-4ae2-b492-5e5dfcc0dc82" />
 </Solution>
